move conservation area getters from load_data to load_geodata
