+++ asf_heat_pump_suitability/getters/load_data.py
@@ -64,6 +64,7 @@ def load_gdf_microsoft_building_footprints(url: str) -> gpd.GeoDataFrame:
     return gdf
 
 
+# TODO move to utils for flats on fossils
 def get_df_osopen_uprn_latlon(**kwargs) -> pl.DataFrame:
     """
     Get raw OS (Ordnance Survey) Open UPRN dataset containing latitude and longitude and British National Grid x and y
@@ -83,41 +84,6 @@ def get_df_osopen_uprn_latlon(**kwargs) -> pl.DataFrame:
     )
 
     return df
-
-
-def load_gdf_historic_england_conservation_areas(**kwargs) -> gpd.GeoDataFrame:
-    """
-    Load GeoDataFrame with building conservation area polygons from Historic England (CRS: EPSG:4326).
-
-    Args:
-        **kwargs for `gpd.read_file()`
-
-    Returns:
-        gpd.GeoDataFrame: polygons of building conservation areas in England
-    """
-    gdf = gpd.read_file(
-        config["data_source"]["E_historic_england_conservation_areas"], **kwargs
-    )
-
-    return gdf
-
-
-def load_gdf_welsh_gov_conservation_areas(**kwargs) -> gpd.GeoDataFrame:
-    """
-    Load GeoDataFrame with building conservation area polygons from the Welsh Government (CRS: EPSG:27700 British
-    National Grid).
-
-    Args:
-        **kwargs for `gpd.read_file()`
-
-    Returns:
-        gpd.GeoDataFrame: polygons of building conservation areas in Wales
-    """
-    gdf = gpd.read_file(
-        config["data_source"]["W_welsh_gov_conservation_areas"], **kwargs
-    )
-
-    return gdf
 
 
 def get_df_ons_number_of_households() -> pl.DataFrame:
