+++ asf_heat_pump_suitability/getters/load_geodata.py
@@ -489,3 +489,38 @@ def load_gdf_listed_buildings(nation: str = "GB", **kwargs) -> gpd.GeoDataFrame:
             "Please set `nation` to either 'England', 'Scotland', 'Wales', or 'GB'."
         )
     return gdf
+
+
+def load_gdf_historic_england_conservation_areas(**kwargs) -> gpd.GeoDataFrame:
+    """
+    Load GeoDataFrame with building conservation area polygons from Historic England (CRS: EPSG:4326).
+
+    Args:
+        **kwargs for `gpd.read_file()`
+
+    Returns:
+        gpd.GeoDataFrame: polygons of building conservation areas in England
+    """
+    gdf = gpd.read_file(
+        config["data_source"]["E_historic_england_conservation_areas"], **kwargs
+    )
+
+    return gdf
+
+
+def load_gdf_welsh_gov_conservation_areas(**kwargs) -> gpd.GeoDataFrame:
+    """
+    Load GeoDataFrame with building conservation area polygons from the Welsh Government (CRS: EPSG:27700 British
+    National Grid).
+
+    Args:
+        **kwargs for `gpd.read_file()`
+
+    Returns:
+        gpd.GeoDataFrame: polygons of building conservation areas in Wales
+    """
+    gdf = gpd.read_file(
+        config["data_source"]["W_welsh_gov_conservation_areas"], **kwargs
+    )
+
+    return gdf
